integrate into test script
